requirements.txt
livekit-agents[openai]>=0.12.0
livekit-api>=0.8.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
